// corp-cargo-main/corp-cargo-main/src/modules/TaxiRegistry/types/types.ts
export type DescriptionsField = [string | JSX.Element, string | string[] | number | number[] | null | undefined];

export type Records = DescriptionsField[];

export interface CoopTripFieldProps { description: string | number | string[] | number[] | null | undefined }

export enum FinishStatusTripFields {
  'Фактическое время ожидания по адресу отправления',
  'Фактическая стоимость, руб',
  'Фактическая дальность, км',
  'Фактическая длительность, мин',
  'Дата внесения фактических параметров поездки',
}

export enum PassengerInfo {
  fullName = 'fullName',
  approvedBy = 'approvedBy',
  department = 'department',
  position = 'position',
  departureAddress = 'departureAddress',
  destinationAddress = 'destinationAddress',
  intermediateAddress = 'intermediateAddress',
  departureWaitTime = 'departureWaitTime',
  destinationWaitTime = 'destinationWaitTime',
  purpose = 'purpose',
  plannedCost = 'plannedCost',
  cost = 'cost',
  plannedDistance = 'plannedDistance',
  distance = 'distance',
  plannedTime = 'plannedTime',
  time = 'time',
  limit = 'limit',
  comment = 'comment',
  rating = 'rating',
  VspGosbTbExist = 'VspGosbTbExist',
  ratingComment = 'ratingComment',
}